refactor(form): move assistance to form account sign in
Resolves #418.
