+++ nuxt/components/form/account/FormAccountSignIn.vue
@@ -32,12 +32,27 @@
         }}
       </AppLink>
     </div>
+    <template slot="assistance">
+      <ButtonColored
+        v-if="
+          graphqlErrorComputed &&
+          graphqlErrorComputed.startsWith('Account not verified!')
+        "
+        :aria-label="$t('verificationMailResend')"
+        @click="accountRegistrationRefresh"
+      >
+        <!-- https://github.com/maevsi/maevsi/issues/209 -->
+        {{ $t('verificationMailResend') }}
+      </ButtonColored>
+    </template>
   </Form>
 </template>
 
 <script lang="ts">
 import { maxLength, minLength, required } from 'vuelidate/lib/validators'
 import { defineComponent } from '@nuxtjs/composition-api'
+
+import ACCOUNT_REGISTRATION_MUTATION_REFRESH from '~/gql/mutation/account/accountRegistrationRefresh.gql'
 import AUTHENTICATE_MUTATION from '~/gql/mutation/account/accountAuthenticate.gql'
 
 const consola = require('consola')
@@ -59,6 +74,19 @@ const FormAccountSignIn = defineComponent({
       graphqlError: undefined,
     }
   },
+  computed: {
+    graphqlErrorComputed(): any {
+      if (!this.graphqlError) {
+        return
+      }
+
+      return [
+        ...((this.graphqlError as any).graphQLErrors?.map(
+          (e: Error) => e.message
+        ) ?? []),
+      ].join(', ')
+    },
+  },
   watch: {
     form: {
       handler(val) {
@@ -70,6 +98,33 @@ const FormAccountSignIn = defineComponent({
     },
   },
   methods: {
+    async accountRegistrationRefresh() {
+      const res = await this.$apollo
+        .mutate({
+          mutation: ACCOUNT_REGISTRATION_MUTATION_REFRESH,
+          variables: {
+            language: this.$i18n.locale,
+            username: this.$v.form.username?.$model,
+          },
+        })
+        .then(({ data }) =>
+          this.$global.getNested(data, 'accountRegistrationRefresh')
+        )
+        .catch((reason) => {
+          this.graphqlErrorInternal = reason
+          consola.error(reason)
+        })
+
+      if (!res) {
+        return
+      }
+
+      this.$swal({
+        icon: 'success',
+        text: this.$t('registrationRefreshSuccess') as string,
+        title: this.$t('sent'),
+      })
+    },
     async submit() {
       try {
         await this.$global.formPreSubmit(this)
@@ -129,11 +184,17 @@ export type FormAccountSignInType = InstanceType<typeof FormAccountSignIn>
 de:
   passwordFound: Passwort wiedergefunden?
   passwordLost: Passwort verloren?
+  registrationRefreshSuccess: Eine neue Willkommensmail ist auf dem Weg zu dir.
+  sent: Gesendet!
   signIn: Anmelden
   username: Nutzername
+  verificationMailResend: Verifizierungsmail erneut senden
 en:
   passwordFound: Password found?
   passwordLost: Password lost?
+  registrationRefreshSuccess: A new welcome email is on its way to you.
+  sent: Sent!
   signIn: Sign in
   username: Username
+  verificationMailResend: Resend verification mail
 </i18n>
